Revert "Mystery Merge"
This reverts commit f747e5875dd2f0581ddb643f4607e6a2be7de738.

+++ Assets/Scripts/PlayerManager.cs
@@ -183,7 +183,7 @@ public class PlayerManager : MonoBehaviour {
         //Deal damage.
     }
 
-    public void TakeDamage(float Damage)
+    public void TakeDamage(int Damage)
     {
         health.CurrentValue -= Damage;
     }
